backend/requirements.txt
fastapi
uvicorn
qdrant-client
psycopg2-binary
sqlalchemy
sentence-transformers
python-multipart
beautifulsoup4
requests
whisper-openai
pydantic
python-dotenv
openai
groq
PyMuPDF